Ship successfully abandoned
Reposting:
When redirecting back to the Webpage (port 8080), the token of the reddit service of port 12129 could no longer be accessed somehow, returning null whenever requesting information from the reddit api. Therefore we decided to ditch that idea on the reddit side of things and to just display the information on a side of the reddit service itself.
The intended version was supposed to work like this, just on the other port/webpage service.